% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{XCNV-package}
\alias{XCNV}
\alias{XCNV-package}
\title{XCNV: Genome-Wide Copy Number Variant Pathogenicity Prediction}
\description{
An R-native interface to the X-CNV algorithm published by Zhang et al.
(2021). The published tree is bundled in a runtime-neutral form; large
annotation resources are supplied by the user. The package also provides an
independently authored, provenance-first implementation of the ACMG/ClinGen
2019 constitutional CNV scoring contract.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rgenomicsetl.github.io/XCNV/}
  \item \url{https://github.com/RGenomicsETL/XCNV}
  \item \url{https://github.com/kbvstmd/XCNV}
  \item Report bugs at \url{https://github.com/RGenomicsETL/XCNV/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkou-bioinfo@users.noreply.github.com}

Authors:
\itemize{
  \item Sounkou Mahamane Toure \email{sounkou-bioinfo@users.noreply.github.com}
}

Other contributors:
\itemize{
  \item Lu Zhang [contributor]
  \item Jing Shi [contributor]
  \item Jing Ouyang [contributor]
  \item Rui Zhang [contributor]
  \item Yan Tao [contributor]
  \item Dong Yuan [contributor]
  \item Cong Lv [contributor]
  \item Rui Wang [contributor]
  \item Bing Ning [contributor]
  \item Richard Roberts [contributor]
  \item William Tong [contributor]
  \item Zhi Liu [contributor]
  \item Tao Shi [contributor]
}

}
\keyword{internal}
