% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{xcnv_bundled_model}
\alias{xcnv_bundled_model}
\title{Return the bundled published X-CNV model}
\usage{
xcnv_bundled_model()
}
\value{
A validated \code{xcnv_portable_model} object.
}
\description{
Return the bundled published X-CNV model
}
